+++ bootique/src/main/java/io/bootique/meta/application/CommandMetadata.java
@@ -1,4 +1,4 @@
-package io.bootique.application;
+package io.bootique.meta.application;
 
 import io.bootique.command.Command;
 import io.bootique.meta.MetadataNode;