+++ bootique/src/main/java/io/bootique/meta/application/OptionValueCardinality.java
@@ -1,4 +1,4 @@
-package io.bootique.application;
+package io.bootique.meta.application;
 
 /**
  * Defines whether an {@link OptionMetadata} takes a value and whether the value is required.