+++ bootique/src/main/java/io/bootique/meta/application/package-info.java
@@ -2,4 +2,4 @@
  * Package contains metadata classes representing a model of the application, its commands and options. All in a
  * format independent from the chosen command-line style.
  */
-package io.bootique.application;
+package io.bootique.meta.application;