+++ bootique/src/test/java/io/bootique/run/DefaultRunnerTest.java
@@ -1,7 +1,7 @@
 package io.bootique.run;
 
-import io.bootique.application.CommandMetadata;
-import io.bootique.application.OptionMetadata;
+import io.bootique.meta.application.CommandMetadata;
+import io.bootique.meta.application.OptionMetadata;
 import io.bootique.cli.Cli;
 import io.bootique.command.Command;
 import io.bootique.command.CommandManager;